update：更新 AppDB 演示案例代码

+++ application/AppDB/src/main/java/afkt/db/database/room/module/note/NoteDao.kt
@@ -57,7 +57,7 @@ interface NoteDao {
     @Transaction
     @Query("SELECT * FROM NoteTable WHERE id > :lastId ORDER BY id ASC LIMIT :pageSize")
     fun getNoteAndPictureListsAfterId(
-        lastId: Int,
+        lastId: Long,
         pageSize: Int
     ): LiveData<List<NoteAndPicture>>
